+++ tests/integration_tests/analysis/snapshots/test_es_update/test_update_only_using_subset_observations/0/update_log
@@ -1,0 +1,17 @@
+======================================================================================================================================================
+Time:
+Parent ensemble: default_0
+Target ensemble: new_ensemble
+Alpha: 3.0
+Global scaling: 1.0
+Standard cutoff: 1e-06
+Run id: id
+Update step: DISABLED_OBSERVATIONS
+------------------------------------------------------------------------------------------------------------------------------------------------------
+                                        Observed history                |                  Simulated data            |      Status
+------------------------------------------------------------------------------------------------------------------------------------------------------
+  1   : FOPR                            0.008 +/- 0.100                 |                 0.079 +/- 0.107            |      Active
+  2   : WPR_DIFF_1                      0.000 +/- 0.100                 |                -0.011 +/- 0.060            |      Active
+  3   : WPR_DIFF_1                      0.100 +/- 0.200                 |                 0.081 +/- 0.126            |      Active
+  4   : WPR_DIFF_1                      0.200 +/- 0.150                 |                 0.073 +/- 0.130            |      Active
+  5   : WPR_DIFF_1                      0.000 +/- 0.050                 |                 0.127 +/- 0.125            |      Active
