+++ tests/integration_tests/cli/snapshots/test_integration_cli/test_es_mda/es_mda_integration_snapshot
@@ -1,0 +1,21 @@
+Iteration,Realization,COEFFS:a,COEFFS:b,COEFFS:c
+iter-0,1,0.703986294605,1.83992056754,4.46053898598
+iter-0,2,0.198965241239,0.881942229016,4.1521713031
+iter-0,4,0.249105100695,1.28419800854,4.90535532003
+iter-0,8,0.558106754935,0.819887823895,2.00307282443
+iter-0,16,0.591623902385,1.7663810255,4.44242760126
+iter-1,1,0.528687318172,1.32403490437,3.69608391642
+iter-1,2,0.446754485528,1.14115565057,3.06814203437
+iter-1,4,0.483635549821,1.57544165709,4.7628689028
+iter-1,8,0.484429573005,0.672512465712,2.05458536861
+iter-1,16,0.389874371197,1.32924399914,4.19237881105
+iter-2,1,0.523645503842,1.19838581664,3.25086460774
+iter-2,2,0.43086689669,0.946942861008,2.36404806355
+iter-2,4,0.463046868611,1.02003340787,3.45834744009
+iter-2,8,0.509481840513,0.618716362006,1.7605474942
+iter-2,16,0.405495589342,0.898454980592,2.70541815914
+iter-3,1,0.533130473238,1.14367178973,2.99509799222
+iter-3,2,0.423337415775,0.964256985466,2.45442264577
+iter-3,4,0.498792171119,1.09932535439,3.54553895903
+iter-3,8,0.542240490526,0.756044439407,2.18736947131
+iter-3,16,0.426694705709,1.13572495324,3.4399530398
