+++ tests/integration_tests/storage/test_field_parameter.py
@@ -1,0 +1,298 @@
+import os
+import stat
+from argparse import ArgumentParser
+from pathlib import Path
+from textwrap import dedent
+
+import numpy as np
+import numpy.testing
+import pytest
+import xtgeo
+
+from ert.__main__ import ert_parser
+from ert.cli import ENSEMBLE_SMOOTHER_MODE
+from ert.cli.main import run_cli
+from ert.config import ErtConfig
+from ert.storage import open_storage
+
+
+@pytest.mark.integration_test
+def test_field_param_update(tmpdir):
+    """
+    This replicates the poly example, only it uses FIELD parameter
+    """
+    with tmpdir.as_cwd():
+        config = dedent(
+            """
+            NUM_REALIZATIONS 5
+            OBS_CONFIG observations
+
+            FIELD MY_PARAM PARAMETER my_param.grdecl INIT_FILES:my_param.grdecl FORWARD_INIT:True
+            GRID MY_EGRID.EGRID
+
+            GEN_DATA MY_RESPONSE RESULT_FILE:gen_data_%d.out REPORT_STEPS:0 INPUT_FORMAT:ASCII
+            INSTALL_JOB poly_eval POLY_EVAL
+            SIMULATION_JOB poly_eval
+        """
+        )
+        with open("config.ert", "w", encoding="utf-8") as fh:
+            fh.writelines(config)
+
+        NCOL = 5
+        NROW = 4
+        NLAY = 1
+        grid = xtgeo.create_box_grid(dimension=(NCOL, NROW, NLAY))
+        grid.to_file("MY_EGRID.EGRID", "egrid")
+
+        with open("forward_model", "w", encoding="utf-8") as f:
+            f.write(
+                dedent(
+                    """#!/usr/bin/env python
+import xtgeo
+import numpy as np
+import os
+
+if __name__ == "__main__":
+    if not os.path.exists("my_param.grdecl"):
+        values = np.random.standard_normal(5*4)
+        with open("my_param.grdecl", "w") as fout:
+            fout.write("MY_PARAM\\n")
+            fout.write(" ".join([str(val) for val in values]) + " /\\n")
+    with open("my_param.grdecl", "r") as fin:
+        for line_nr, line in enumerate(fin):
+            if line_nr == 1:
+                a, b, c, *_ = line.split()
+
+    output = [float(a) * x**2 + float(b) * x + float(c) for x in range(10)]
+    with open("gen_data_0.out", "w", encoding="utf-8") as f:
+        f.write("\\n".join(map(str, output)))
+        """
+                )
+            )
+        os.chmod(
+            "forward_model",
+            os.stat("forward_model").st_mode
+            | stat.S_IXUSR
+            | stat.S_IXGRP
+            | stat.S_IXOTH,
+        )
+        with open("POLY_EVAL", "w", encoding="utf-8") as fout:
+            fout.write("EXECUTABLE forward_model")
+        with open("observations", "w", encoding="utf-8") as fout:
+            fout.write(
+                dedent(
+                    """
+            GENERAL_OBSERVATION MY_OBS {
+                DATA       = MY_RESPONSE;
+                INDEX_LIST = 0,2,4,6,8;
+                RESTART    = 0;
+                OBS_FILE   = obs.txt;
+            };"""
+                )
+            )
+
+        with open("obs.txt", "w", encoding="utf-8") as fobs:
+            fobs.write(
+                dedent(
+                    """
+            2.1457049781272213 0.6
+            8.769219841380755 1.4
+            12.388014786122742 3.0
+            25.600464531354252 5.4
+            42.35204755970952 8.6"""
+                )
+            )
+
+        parser = ArgumentParser(prog="test_main")
+        parsed = ert_parser(
+            parser,
+            [
+                ENSEMBLE_SMOOTHER_MODE,
+                "--current-case",
+                "prior",
+                "--target-case",
+                "smoother_update",
+                "config.ert",
+            ],
+        )
+
+        run_cli(parsed)
+        config = ErtConfig.from_file("config.ert")
+        with open_storage(config.ens_path, mode="w") as storage:
+            prior = storage.get_ensemble_by_name("prior")
+            posterior = storage.get_ensemble_by_name("smoother_update")
+
+        prior_result = prior.load_parameters("MY_PARAM", list(range(5)))
+        assert len(prior_result.x) == NCOL
+        assert len(prior_result.y) == NROW
+        assert len(prior_result.z) == NLAY
+
+        posterior_result = posterior.load_parameters("MY_PARAM", list(range(5)))["values"]
+        # Only assert on the first three rows, as there are only three parameters,
+        # a, b and c, the rest have no correlation to the results.
+        assert np.linalg.det(
+            np.cov(prior_result.values.reshape(5, NCOL * NROW * NLAY).T[:3])
+        ) > np.linalg.det(
+            np.cov(posterior_result.values.reshape(5, NCOL * NROW * NLAY).T[:3])
+        )
+        # This checks that the fields in the runpath are different between iterations
+        assert Path("simulations/realization-0/iter-0/my_param.grdecl").read_text(
+            encoding="utf-8"
+        ) != Path("simulations/realization-0/iter-1/my_param.grdecl").read_text(
+            encoding="utf-8"
+        )
+
+
+@pytest.mark.integration_test
+def test_parameter_update_with_inactive_cells_xtgeo_grdecl(
+    tmpdir
+):
+    """
+    This replicates the poly example, only it uses FIELD parameter
+    """
+    with tmpdir.as_cwd():
+        config = dedent(
+            """
+            NUM_REALIZATIONS 5
+            OBS_CONFIG observations
+            FIELD MY_PARAM PARAMETER my_param.grdecl INIT_FILES:my_param.grdecl FORWARD_INIT:True
+            GRID MY_EGRID.EGRID
+            GEN_DATA MY_RESPONSE RESULT_FILE:gen_data_%d.out REPORT_STEPS:0 INPUT_FORMAT:ASCII
+            INSTALL_JOB poly_eval POLY_EVAL
+            SIMULATION_JOB poly_eval
+        """
+        )
+        with open("config.ert", "w", encoding="utf-8") as fh:
+            fh.writelines(config)
+
+        NCOL = 4
+        NROW = 4
+        NLAY = 1
+        grid = xtgeo.create_box_grid(dimension=(NCOL, NROW, NLAY))
+        mask = grid.get_actnum()
+        mask_list = [True] * 3 + [False] * 12 + [True]
+        mask.values = mask_list
+        grid.set_actnum(mask)
+        grid.to_file("MY_EGRID.EGRID", "egrid")
+
+        with open("forward_model", "w", encoding="utf-8") as f:
+            f.write(
+                dedent(
+                    """#!/usr/bin/env python
+import xtgeo
+import numpy as np
+import os
+if __name__ == "__main__":
+    if not os.path.exists("my_param.grdecl"):
+        values = np.random.standard_normal(4*4)
+        with open("my_param.grdecl", "w") as fout:
+            fout.write("MY_PARAM\\n")
+            fout.write(" ".join([str(val) for val in values]) + " /\\n")
+    with open("my_param.grdecl", "r") as fin:
+        for line_nr, line in enumerate(fin):
+            if line_nr == 1:
+                a, b, c, *_ = line.split()
+    output = [float(a) * x**2 + float(b) * x + float(c) for x in range(10)]
+    with open("gen_data_0.out", "w", encoding="utf-8") as f:
+        f.write("\\n".join(map(str, output)))
+        """
+                )
+            )
+        os.chmod(
+            "forward_model",
+            os.stat("forward_model").st_mode
+            | stat.S_IXUSR
+            | stat.S_IXGRP
+            | stat.S_IXOTH,
+        )
+        with open("POLY_EVAL", "w", encoding="utf-8") as fout:
+            fout.write("EXECUTABLE forward_model")
+        with open("observations", "w", encoding="utf-8") as fout:
+            fout.write(
+                dedent(
+                    """
+            GENERAL_OBSERVATION MY_OBS {
+                DATA       = MY_RESPONSE;
+                INDEX_LIST = 0,2,4,6,8;
+                RESTART    = 0;
+                OBS_FILE   = obs.txt;
+            };"""
+                )
+            )
+
+        with open("obs.txt", "w", encoding="utf-8") as fobs:
+            fobs.write(
+                dedent(
+                    """
+            2.1457049781272213 0.6
+            8.769219841380755 1.4
+            12.388014786122742 3.0
+            25.600464531354252 5.4
+            42.35204755970952 8.6"""
+                )
+            )
+
+        parser = ArgumentParser(prog="test_main")
+        parsed = ert_parser(
+            parser,
+            [
+                ENSEMBLE_SMOOTHER_MODE,
+                "--current-case",
+                "prior",
+                "--target-case",
+                "smoother_update",
+                "config.ert",
+            ],
+        )
+
+        run_cli(parsed)
+        config = ErtConfig.from_file("config.ert")
+        with open_storage(config.ens_path) as storage:
+            prior = storage.get_ensemble_by_name("prior")
+            posterior = storage.get_ensemble_by_name("smoother_update")
+
+            prior_result = prior.load_parameters("MY_PARAM", list(range(5)))
+            posterior_result = posterior.load_parameters("MY_PARAM", list(range(5)))["values"]
+
+            # check the shape of internal data used in the update
+            assert prior_result.shape == (5, NCOL, NROW, NLAY)
+            assert posterior_result.shape == (5, NCOL, NROW, NLAY)
+
+            # Only assert on the first three rows, as there are only three parameters,
+            # a, b and c, the rest have no correlation to the results.
+            assert np.linalg.det(
+                np.cov(prior_result.values.reshape(5, NCOL * NROW * NLAY).T[:3])
+            ) > np.linalg.det(
+                np.cov(posterior_result.values.reshape(5, NCOL * NROW * NLAY).T[:3])
+            )
+
+            # This checks that the fields in the runpath
+            # are different between iterations
+            assert Path("simulations/realization-0/iter-0/my_param.grdecl").read_text(
+                encoding="utf-8"
+            ) != Path("simulations/realization-0/iter-1/my_param.grdecl").read_text(
+                encoding="utf-8"
+            )
+
+            # check shapre of written data
+            prop0 = xtgeo.grid_property.gridproperty_from_file(
+                "simulations/realization-0/iter-0/my_param.grdecl",
+                fformat="grdecl",
+                grid=grid,
+                name="MY_PARAM",
+            )
+            assert len(prop0.get_npvalues1d()) == 16
+            numpy.testing.assert_array_equal(
+                np.logical_not(prop0.values1d.mask), mask_list
+            )
+
+            prop1 = xtgeo.grid_property.gridproperty_from_file(
+                "simulations/realization-0/iter-0/my_param.grdecl",
+                fformat="grdecl",
+                grid=grid,
+                name="MY_PARAM",
+            )
+            assert len(prop1.get_npvalues1d()) == 16
+            numpy.testing.assert_array_equal(
+                np.logical_not(prop1.values1d.mask), mask_list
+            )
